bug fix: UI_Markdown kept updating size, causing flashing
when the same markdown was sent to UI_Markdown, it would cause a the
element to become dirty and need its size recomputed.  This would cause
the element to change sizes frequently, so the ui would flash.  now, the
markdown is cached, so recalcs are prevented if the same markdown is
set.